Plotter, PID Controllers
Have added a plottter to the PID class. Added trackbars for gain
constants. For some reason I cannot create more than one instance of the
class. I think it has something to do with the threads, as when I do NOT
create a thread to run the GUI inside the PID class, it creates all 3
objects fine. It is only when I try to create more than 1 threads inside
a class.